Cleanup, Brake Mode, & Debugging.
So cleaned up a tiny bit of the code from stuff that we dont use
anymore. Jason wanted me to test a way to test the ability to switch
between brake and coast mode during the math so that way we can stay on
the ramp at the end. Finally, fixed an possible issue with null pointers
preventing the autonomous code from running.