Rearranged controls
It just looked like a jumble so I have rearranged it, I tried committing this yesterday and it wouldn't go through so I am trying it again. I rearranged the mapping and controls in OI.cpp so it's in opJoystick and driverjoystick categories-I think hopefully it will be easier to understand now, as well as see which buttons are being used or not used, which I have in a google doc if anyone needs to see it, but you could find the same info on here. Also an FYI there are 13 "buttons" total that include pressing down the joysticks, the select button, the start button, and the home button.